LWSHADOOP-762: Update `pig-solr` to Pig 0.17.0 and Hadoop 3
With this commit, `pig-solr` now builds with the latest version of Pig,
and a much more recent Hadoop release (Hadoop 3.0.0).

There is still one last vestige of Hadoop 2.  A few tests use a
test package called "MRUnit".  MRUnit is a dead project, and won't be
moving to Hadoop 3.  We'll need to eventually replace our use of MRUnit
to get these last traces of the hadoop2 jars off of the test classpath.